requirements.txt
scrapy
newsplease
google-cloud-datastore